Update creative-brief.html
Here I've updated and completed the rest of our creative brief HTML. We can still make small edits to this if anyone sees anything that should to be changed or added, as we still have some time until it's due Sunday, but for the most part this is finished. The css file is also complete.