Update to creative-brief.html
I've included all the updates that Lydia previously suggested. If anyone sees any other changes that need to be made let me know or feel free to edit yourself. Lydia, I think you said you saw a few grammar mistakes so if you want to edit this file to fix them that's completely fine with me. Also note that before we submit all our files to the professor, we should probably check to be sure that the "SWOT Analysis" hyperlink is working on this page.