First version of service banner component (.html, .css, .jpg)
This is my first draft of the service banner component html and css files, along with the image that I'm using for the time being. The colors and fonts are not final. There are some things I want to tweak before this is all due Sunday but let me know what you all think!