Update index.html
This is an update to the service banner component HTML file. I incorporated the grid system and made some other changes. I'm having an issue with the image shrinking too small in tablet mode. For a second, the image lifts up and doesn't look proportionate to the rest of the text. If you have any suggestions on how that could be fixed feel free to let me know, will finish it up tomorrow.